20210315 임시저장

+++ Prs_level01/src/com/level01/getAverage.java
@@ -19,7 +19,7 @@ public class getAverage {
 		}
         
         
-        answer = sum / arr.length;
+        answer = sum / (double) arr.length;
         System.out.println(answer);
         return answer;
     }
